refactor: extract sanitize_for_log to shared src/util.rs module
`sanitize_for_log` was added to `mcp/server.rs` in PR #154 as a
private helper for the `initialize` handler (escape control chars +
cap length on client-controlled strings before logging). The same
hardening pattern is needed in `git2_ops/push.rs::unbundle`, where
git's stderr (potentially containing ANSI escapes, newlines, or
oversized output from a maliciously-crafted bundle) currently flows
unsanitised into our error messages.

Rather than duplicate the helper, extracted it to a new
`src/util.rs` module with `pub fn sanitize_for_log` and the
`MAX_LOG_STRING_LEN` constant (renamed from
`MAX_CLIENT_STRING_LOG_LEN` since the helper is now generic, not
client-specific). Re-exported via `pub mod util` in `lib.rs`.

Updated `mcp/server.rs` to import from the shared module; removed
the local copy + its 4 tests (which moved to `util.rs::tests` —
plus 3 additional tests for completeness: NUL-as-short-form,
unicode preservation, and ANSI-escape isolation).

Pure refactor: no behaviour change. Same `escape_debug` semantics,
same 200-byte cap, same `…` truncation marker, same UTF-8-boundary-
safe truncation. Sets up commit 3 to apply the same hardening to
`unbundle`'s stderr handling.

7 util-module tests + existing 56 server tests + 7 bundle tests
all pass. fmt + clippy strict clean.
